Changed types of pointers to memory

+++ src/kernel/fork.c
@@ -4,7 +4,7 @@
 #include "scheduler.h"
 #include "debug_printf.h"
 
-int8_t copy_process(uint64_t function, uint64_t arg) {
+int8_t copy_process(uintptr_t function, uintptr_t arg) {
     preempt_disable();
     struct task *new_task;
     struct task *current = get_current_task();
@@ -19,9 +19,9 @@ int8_t copy_process(uint64_t function, uint64_t arg) {
 
     new_task->cpu_context.x19 = function;
     new_task->cpu_context.x20 = arg;
-    new_task->cpu_context.pc = (uint64_t) start_coroutine;
+    new_task->cpu_context.pc = (uintptr_t) start_coroutine;
 
-    new_task->cpu_context.sp = (uint64_t) new_task + 4096;
+    new_task->cpu_context.sp = (uintptr_t) new_task + 4096;
 
     add_task(new_task);
     preempt_enable();
