+++ src/kernel/include/fork.h
@@ -7,6 +7,6 @@
 * Copies current process context and starts a new one
 * with the specified argument
 */
-int8_t copy_process(uint64_t function, uint64_t arg);
+int8_t copy_process(uintptr_t function, uintptr_t arg);
 
 #endif
